backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
django-cors-headers>=4.3
psycopg2-binary>=2.9
xgboost>=2.0
scikit-learn>=1.3
pandas>=2.1
numpy>=1.26
joblib>=1.3
requests>=2.31
python-dotenv>=1.0
gunicorn>=21.2
